Copy files over from fboss/agent
Summary:
I'm leaving the files that are needed in both places (fboss/qgent and fboss/qsfp_service) so that the two aren't coupled. In other words we can deploy the service without having to make the agent changes at the same time. Then we can change the agent to rely on this service once it's fully deployed, and so on.

The only changes to these files is changing the incldue locations to include from qsfp_service rather than the agent folder.

Reviewed By: aeckert

Differential Revision: D3868804

fbshipit-source-id: 35fd4f9aa1868f4ea04da17e022727a3b8732308