Remove resetting of CP2112 device in CP2112::open(bool)
Summary:
This diff reverts the modifications made to CP2112::open(bool) in D2970336.

More specifically, D2970336 copied a subset of the logic in CP2112::resetDevice() into CP2112::open(bool), and wrapped it with a resetDevice() if it threw a LibusbError. This is superfluous because this logic already exists in CP2112::resetDevice(), which is called immediately after CP2112::open(bool) in BaseWedgeI2CBus::open().

Differential Revision: D3941824

fbshipit-source-id: 8b9ccd9e724f48938b43fe3aec65679861af03a6