Add CR4, CAUI mode support and enable FEC for 100G ports in lldp_tool
Summary:
i) CR4 is needed for both 6packs and Galaxy. We are already
using this in LLDP configs for 6pack and Galaxy lldp_tool in
ramdisk. However in lldp_tool we only had code for KR4 mode.
For 40G both (default) SR4 and CR4 would do the job, so we
didn't notice the problem, but for 100G we need CR4.
ii) For 100G ports we need to enable/disable FEC on both sides
link for the link to be functional. So to be able to discover
lldp neighbors in ramdisk do this in lldp_tool as well.
The caveat here is that for Galaxy at least, we theoretically
don't actually need to set FEC on the backplane ports.
This is controlled via configs whereby for backplane ports
we set disableFEC to true.
iii) Add CAUI mode support so if we go to 100G DCs where we
have actual 100G front panel ports we don't have a mode
mismatch with the uplink ports.
iv) Get rid of the code for KR4 mode, its unlikely we will
ever use it
v) Sync Lldp thrift from D3929065

Reviewed By: aeckert

Differential Revision: D3925104

fbshipit-source-id: 5dd5d4b013191eb796e743a52edbe71817d5ab84