Add platform specific managers
Summary: This moves across some basic info - at the moment just the bus and the number of transceivers the platform should have. I played around with putting the portmapping in this service as well, but figured it might be simpler if the service just works off transceiver number, and fboss agent worries about how transceivers are mapped to its port structure.

Reviewed By: aeckert

Differential Revision: D3929241

fbshipit-source-id: 96d9ca6242a1bfc5c1a1356ba2f21d81b13e7e1b