Set default values for transceiver tech so mimic old behaviour + explicit handling for sixpack backplane
Summary: Given cp2112 flakiness, we often don't get the correct transceiver type, so set explicit defaults as we set 40G to optical, with all others being copper. This also means we need to explicitly handle the 40G sixpack backports, which won't have this information, but should be set to copper. Mmmm, business logic.

Reviewed By: aeckert

Differential Revision: D3952258

fbshipit-source-id: d68b0769a0e924e5c6b23c6fb01d2cf05a5e6d4f