Fix lingering v6 ip rules in linux
Summary:
The rules and addresses are programmed hand-in-hand by TunManager however Linux
treats them differently.

We were declaring all of our interfaces as DOWN when we first discovers them
from linux. And then we were learning addresses from it. IPv6 addresses are not
learned if interface status is down (for front0). And then we delete front0,
which wipes off all routes and addresses associated with it but not rules.

Lingering rules may cause us some trouble if they are of higher priority. Better
to clean them up well.

Differential Revision: D3965406

fbshipit-source-id: 41c41cd4957089b5ae94fa015c2783852330c00c