Keep track of # of ECMP groups we have programmed and report stats
Summary:
In the host table when adding or removing a ecmp
egress, inc/dec a counter to keep track of number of ECMP
object programmed. Ideally BCM would provide a api to
get the current count of ECMP entries, but the closest they
have for this is ECMP traversal, which is inefficient.

Differential Revision: D4012615

fbshipit-source-id: 47440982b04dd85634611403ab8a5c156c4bc43a