cp2112: don't loop in resetDevice
Summary:
In CP2112::resetDevice, if device is not reachable, we end up looping for a
whole second.  If, in addition, we do this for all transceivers, we end up
looping for as many transceivers as there are, which on wedge40 could be 16.
This is a problem because we call detectTransceiver in FunctionScheduler
every one second, with unexpected consequence that each call can take up to
16 seconds.

With this fix, we will reset the device and only try to close/open it once.

Differential Revision: D4101307

fbshipit-source-id: 1416067995abfc210fa3e7a5989c4d4e2dd759d0