Unittests for routing of packets between Host and Switch
Summary:
As titled. Our code somehow magically works for routing of packets between CPU
and Host and it didn't had any unittest. Further we had many recent changes to
enable link-local and multicast routing (diffs in review).

Adding some exhaustive suite of unittests for unicast, multicast and link-local
routing (for v4 as well v6) of packets between Host and Switch.

We already have a MockPlatform which can be used to test packet being sent to
hardware. I created MockTunManager which is used to test packets flowing to
Host from Agent.

Reviewed By: plapukhov

Differential Revision: D3844049

fbshipit-source-id: 27dccef67d4489d51256229309214be0a2e49ef5