Fix `getPortStatus` thrift call
Summary:
I have also seen crash with `getPortStatus` call and it has been fixed in this
diff. Very similar issue like D4196983.

Also there is a huge regression in thrift call completion timeout because of
transciever stuff. We should avoid exposing that information with port-status
which is meant to be light weight and expose it into a separate thrift API
I have commented out in this diff on purpose to show perf improvement. But I
happy to keep it if you think it is must needed off the box. We should
fix why it takes so long to get transceiver info in a seperate diff.

Differential Revision: D4197145

fbshipit-source-id: 9499f68eaf740bc99e48a81284b5c4bbad408213