Add missed-out commit directly to github
Summary:
Push to github has been failing since Oct 3, 2016.  On investigation
it appears that a commit from January 2016 was not pushed out, which
is resulting in merge issues because of a modification to one of
the files that the Oct commit is changing.  We will resolve this issue
by doing this commit from github.

Reviewed By: has