fix clearing low power mode
Summary:
This fixes a bug where we are not clearing low power mode
correctly after a full chassis power reset. The issue was that the
default value is not 0x2, but is actually set to 0x0. Our logic was,
'if set to 0x2, clear low power if supported'. This changes it to 'if
0x0, clear low power if supported'.

The docs aren't particularly clear about the default. 0x2 means that
low power is set, but is controlled via software. If the value is 0x0,
it means that the power class is controlled via the lpmode pin, which
is the default. On wedge40/sixpack this pin is not connected so we
need to set the power class via software.

Reviewed By: saifhhasan

Differential Revision: D4203293

fbshipit-source-id: aea957cb2d61554ad34b739ea94df99da5f16523