(codegen) mocking wrappers for BCM API calls
Summary:
This attempts to create basic wrappers around BCM SDK calls (OpenNSL and BCM). It does by simply replacing symbols at compile time using the --wrap directive (meaning all symbols needs to be listed in TARGETS). For now, I used some quick and dirty hacks to find the bcm_|sal_|soc_ symbols and generate wrappers for them. For OpenNSL, simply enumerated all function signatures in the header files. The #ghetto tools I used are pretty ad-hoc, I hope we can just add new wrapped methods as we move on.

The wrappers redirect calls to GMOCK instances and we use the linker to replace original symbols with wrappers when building tests. There is one SDK mock instance per thread (I'm not sure this will be really needed, but did it just in case), and every test is required to re-set mock object to start clean, and explicitly destory it (otherwise GMOCK whines about non-destroyed object). Basically, every test is requried to do the following:

- reset the mock instances (so that wrappers can access it)
- destory mock instances (GMOCK checks that the instance has been destructed)

This is not very useful on its own, at the moment. The idea is to combine it with other mocked objects (e.g. BcmSwitch) to do meaningful unit testing of BcmEgress, for example. By no means this allows for stubbing the SDK, but it adds a bit more options to testing.

Reviewed By: saifhhasan

Differential Revision: D4209316

fbshipit-source-id: cf8b63fe811434527ea1e8196454533c874d7949