remove CHECK from getForwardInfo. audit callsites for isResolved
Summary:
There are a few thrift endpoints that naively call getForwardInfo
that can blow up the whole agent. In those cases it is nicer to just
act like we don't have a route. To prevent future fiascos, take out the
CHECK. All the other callsites already explicitly check isResolved and
skip. I added the CHECK to the programming and arp/ndp paths which seem to
legitimately hit it. Not sure if it is preferred to log and skip there too..

Differential Revision: D4224049

fbshipit-source-id: 88d7abf624a436a674c9d9354c9f0440c077daea