Change service to fire preview dapper jobs
Summary:
This changes the main upgrade service to create preview
dapper jobs and fire them instead of directly simulating prefixes.

This has a few benefits:

1. Certain preflight handlers don't work correctly when not run as
part of a maintenance.
2. Saves a ton of time in the service spent simulating preflights.

The algorithm is changed as well: we find a set of racks that would
not violate any service atonce guarantees and start preview jobs on
them. We then watch to see if any of the jobs succeed. If they do, we
then start the normal upgrade, passing in the maintenance. The upgrade
will then push the drain through, perform the upgrade and then
undrain.

I also spent some time optimizing some slow points in the service by
adding some parallelism, taking advantage of using gevent.

Differential Revision: D4203526

fbshipit-source-id: 489e3cce03d727c2928e9706c11d21ba4662c607