Change getTransceiverInfo call to call new QsfpService instead of agent
Summary:
!!! I won't land this until the QsfpService is deployed everywhere !!!

Changing so that this call no longer goes to the agent. The list_optics call didn't work at all before - threw an out of range exception - so fixed that to work with the new service.
It also just prints info for all the transceivers, not just the optical ones? Either way, haven't changed any logic other than where the thrift call is made to.

Differential Revision: D3986588

fbshipit-source-id: 4f0fb3e380cf57a239be601155c0972e413464da