Rework the way ports and transceivers are mapped
Summary:
This leaves the logic for Wedge100 and Galaxy the same, but changes how Wedge40 is determined to mimic the other two (with a static mapping).
It also makes sure that each Port knows what Transceiver and Channel it maps to (useful for later accounting).
In Wedge40, I'm guessing there are some 6pack settings included in the same class. While the backplane ports don't have transceivers (or channels I guess?),
do the extra 16 ports for a LC? Are they quads as well?

Reviewed By: saifhhasan

Differential Revision: D4173446

fbshipit-source-id: 84b1eaffec552e8f9e41cfb026f2ae02e7369b75