Get rid of isPortUp from SwSwitch
Summary:
This is a redundant call at most places. Earlier we didn't had `operState` of
a port (from HW) in `Port` state. I recently added it as a part of SwitchState
to reflect it in the linux.

With this change we are going to save lot of calls on `getState()` which
acquires the global lock to copy state pointer.

Reviewed By: ninas

Differential Revision: D4196593

fbshipit-source-id: aea49fd72efdfbc4544dde79463bb1ad4f761188