Qsfp service only
Summary:
I need to separate the qsfp and agent code so that deployment isn't a nightmare. Easiest way seems to be to squash all commits and then use hg split (sadly can only use this on head).
This commit includes the following changes to the qsfp_service
1.
This moves the call to customize when the link speed changes. It seems to make sense that most of the communication between the two services will happen via WedgePort. If there's somewhere people feel is better, let me know.
D4227803
2.
[fbossagent] Call qsfp_service to get TransmitterTechnology
This moves this logic to the qsfp_service (and is the last thing to be moved \o/) At the moment there is onyl special casing for sixback backplane ports. Is similar needed for anything else?
D4147804
3.
[fbossagent] Call qsfp_service to get TransmitterTechnology
Splitting the qsfp_service stuff from fboss agent to make deployment easier
D4227890

Reviewed By: saifhhasan

Differential Revision: D4228001

fbshipit-source-id: ae2c90427050e42b3d69a721ffeb879f7c21d651