Add simple unit-tests for BcmEgress class
Summary:
This is a slightly more elaborte example of using the OpenNSL mock functions to test BcmEgress object. For now it tests the paths that do not utilize the WarmBootCache.

The tests are rather trivial, but illustrate how you can combined mocked SDK with other Mocks. Among the changes made to the main code is replacing BcmSwitch with BcmSwitchIf so we can easily add mocked version of the latter, without inheriting state/constructors and all that jazz.

This is not intended to be comprehensive, but rather illustrates how basic unittests could be done while mocking SDK responses.

Reviewed By: saifhhasan

Differential Revision: D4218574

fbshipit-source-id: 83ef36079ac977a5d424927901b204a8cbb14ed0