Add wedge_agent stub-SDK target (fboss emulator)
Summary:
This adds basic L3 switching functionality to the SDK stub. In current version, it adds new target "wedge_agent-stubbed" that links with CombinedSdk and has many of the SDK calls stubbed to give perception of a virtual ASIC. The modeled platform is wedge with 128 ports (speed is irrelevant, not checked).

Under the hood, we model the basic BRCM object tables: egress objects, interface objects, ECMP objects, L3 Host, L3 Route tables. There is also bunch of hack to get FBOSS to initialize.

This version supports basic routing using lookup in Hosts and L3 Route tables. It does ECMP switching for v4/v6 packets based on flow hash, TTL decerment and all that. It does also local multicast switching, and LLDP propagation.

We rely on FBOSS process supplying VLAN tag to properly identify the outgoing port for non-L3 traffic. The tag is preserved in all packets as they are being switched through the message hub, but only used on the first switch in the chain.

Caveats:

- Warm boot not supported (disabled explicitly on start)
- MAC address needs to specified per interfacec and be unique - we need unique link-local addresses per port (due to way we handle host table)
- Port scheme is fixed to 128 ports
- Punting to CPU is done in pretty ad-hoc fashion, we do not use FP tables currently

Differential Revision: D4387918

fbshipit-source-id: 90caf1be3360cab1a991980bd94cb99445166b14