Make some BcmTests work with emulated SDK
Summary:
this is a demo of how emulated SDK could be used to build unit-tests that involved SDK interaction. For starters, I took the BcmTests that we previously ran on real hardware. I currently only got interface/route tests working, adding the missing parts in the SDK.

One main reason for not (yet) making this a cpp_unittest is singleton failure - BcmUnit singleton is destructed after the SDK singleton, which caused SIGSEGV on test termination. I have a plan to fix that, but it would require significant lines of changes (replacing try_get()->XXX with actual pointer validation) across the whole SDK.

TODO: Make VLAN/Port tests work as well

Differential Revision: D4420344

fbshipit-source-id: 92775e58057a56aff0420831888ee653b4476fb5