Always get SN from fbwhoami
Summary:
With t14932164, we are now creating virtual assets for
Fabric cards on Galaxy as well. These cards get 'F' appended to
their SN.  Using the existing logic, we could always append 'F'
for Galaxy FABs. Unfortunately we have 14 chassis which were
provisioned with the earlier logic of not creating virtual assets
for FCs.
So, just simplify the logic and always use fbwhoami to get the SN.
We use fbwhoami anyways for the case where we rest-api fails.
There earlier concern about this was regarding, serf (and
thus fbwhoami having incorrect data). However this seems unfounded,
is SN in serf is wrong, we will fail in provisioning itself
and never get to running FBOSS agent.

Reviewed By: saifhhasan

Differential Revision: D4440689

fbshipit-source-id: 7900f2ef3955e40e59c25dc289025886d34c0c5e