Fix lldp -vv in fboss cli
Summary:
FBOSS agent returns binary data for raw chassis/port id.
This is equivalent to array of bytes in python. In py2
string = array of bytes, not so in py3. So we can't iterate
over array of bytes and consider each byte as a 1 character
substring and then check if that characters is in array of
printable chars. Instead for py3 just check if this can
be decoded as a utf-8 string.

Reviewed By: starskywongfb

Differential Revision: D4549194

fbshipit-source-id: 4580e69225a1c1df0fc40f4be87fb8cbdcdfae4b