Fix BcmSwitch destruction to be graceful
Summary:
Since we never actually destroy BcmSwitch in production, the
destructor has some obvious crashes if we try to run the
destructor. One is that we try to unregister bcm callbacks multiple
times. Another is some dependent member vars being destroyed out of
order. This is needed to start running unit tests on actual hw.

Reviewed By: ninas

Differential Revision: D4565685

fbshipit-source-id: 5491b0d1a3e2b938c20ba07ef26c6329b28ce4f6