clean up WedgePlatforms
Summary:
This refactors our Platform (+PlatformPort) code to expose a
uniform interface and to remove duplicate or unused code. The main
change is factoring the PortMapping logic in to a helper class. That
class is responsible for instantiating the port map properly for
different platforms, given a mapping from logical port -> frontpanel
port. We had a lot of duplicate code for implementing initPorts on
different platforms.

I also split any Platforms that used a "mode" variable to support
multiple platforms in the same class. Wedge40Platform supported
Wedge40, SixpackLC and SixpackFC and had different logic based on the
passed in mode. GalaxyPlatform supported both GalaxyLC and
GalaxyFC. There are now specific platform classes for these. This may
only differ in the portmap, but it is much easier to reason about and
now every platform is constructed w/ only one parameter--the
productInfo.

Reviewed By: ninas

Differential Revision: D4566335

fbshipit-source-id: 5e854b81fe98dec4aec867c939595f38643301a0