Add port remediater flow that will recall customizeTransceinver on ports that are down
Summary:
This retriggers the setSpeed flow for down ports. It won't actually flap the port unless the port is already down. It uses the current expected speed to reconfigure the qsfps for that port.
This is to solve the issue where a transceiver is unplugged (thereby wiping settings) and then plugged back in. The agent registers port down events, but not port back up ones (because the port won't go back up without the correct qsfp settings).
It was decided that doing this in the agent is nicer than qsfp service, because this way the service doesn't have to hold any state.

Reviewed By: aeckert

Differential Revision: D4588488

fbshipit-source-id: ea9ff4cdb8c8359edde3c25184094e1a43a28175