Only customize once for each transceiver using the first channel speed
Summary: Currently we customize for each channel, meaning that the speed of the last channel wins, meaning for 100G we pretty much always set the transceiver to 25G settings. This is...not ideal. This diff changes it to use the speed of the controlling port which is the first channel.

Reviewed By: aeckert

Differential Revision: D4605018

fbshipit-source-id: 6b98daa7cd6cb0852e9ee06f735a8cbeb5ed66db