Create Mockables for testing.
Summary:
This creates two new classes, MockablePlatform and
MockableHwSwitch. These expose the same mock interface as MockPlatform
and MockHwSwitch, but by default delegate mock calls to a real
HwSwitch/Platform implementation. This means that if the unit tests
are running on actual hardware we can pass in an initialized BcmSwitch
or WedgePlatform to exercise the code in each of those classes. This
diff does not actual hook up a real platform in any of our tests (done
in a later diff).

Reviewed By: ninas

Differential Revision: D4565664

fbshipit-source-id: 54fa7bea454a5d51610e339e4604a15f135fc9f0