Allow running unit tests on actual hw
Summary:
This updates TestUtils and our various tests to optionally
exercise our real BcmSwitch/WedgePlatform code if on switch
hardware. This is currently enabled via a gflag, though I will think
about a nicer way of enabling this. If --switch_hw is passed in, we
exercise the full initialization of a switch (starting w/
WedgeProductInfo), run the test, then fully destroy the switch.

For now, the tests don't exercise much beyond this, but this is a
start. I had to modify some tests because programming an empty state
crashes BcmSwitch as it tries to delete the default vlan (1).

Plan going forward is to add a handle class that makes it easy to
trigger real packet receipts or link events in tests and then also add
more intensive tests that take advantage of this mechanism (warm boot
tests, ecmp shrink/expand etc.).

Differential Revision: D4565749

fbshipit-source-id: 30a8a6dee90ec45eff634197587373ada6e1c39b