make the default MockHwSwitch a NiceMock
Summary:
My recent diff adding Mockables causes us to mock a lot more
functions in HwSwitch. I tried to use NiceMock everywhere to avoid the
warnings for unnecessary calls, but missed the default case (not
running with --switch_hw). Hopefully we can remove this down the road
when we lock down possible inputs, but this at least makes test output
readable again in the meantime.

Differential Revision: D4701342

fbshipit-source-id: 123f5d0099d36891a07b7aff8cbf3a3740b9b3c0