Don't crash on link events if port is not in port group
Summary:
It is apparently possible on sdk 6.4.6 that ports are not
actually registered in port groups unless you pass in the --flexports
flag. This is because the bcm_get_subsidiary_ports call return
BCM_E_PORT. Luckily, we do not run with this configuration anywhere
outside of a handful of wedge100s that should be upgraded, but let's
still not crash in this case. Noticed this running unit tests on
actual switch hw.

Differential Revision: D4701404

fbshipit-source-id: ad4170373427d90595685d46c42c3fb918582c79