Add counter for v4/v6 header errors. Trap dest miss packets to CPU
Summary:
i) Add counters for v4/v6 header errors. These would get incremented in
case of
a) Malformed packets
b) Packets with TTL = 0 showing up.
In practice these counters should always be 0.
ii) Trap packets for which we can't route to CPU low pri queue -
Motivation here is to be able to see a subset of packets are being
dropped due to no route being available for them. In our n/w, we
should always have a default route, so packets getting trapped to
us for this reason would point to a serious BGP bug.
A follow up action is to increment a counter for such destination
lookup failure cases

Reviewed By: saifhhasan

Differential Revision: D4711558

fbshipit-source-id: 8dd18b4ef94b2544df6e2eadaf3234c9186e48b4