change ports w/out config to be disabled (not down)
Summary:
The logic for applying a thrift config sets any already
configured ports to be DOWN if they don't exist in the config. Now
that we reflect port state in to the SwitchState, this causes an
inconsistency where the port is actually up, but is down in the
SwitchState. To avoid this, let's just disable any ports in hw, but
not in config. This *should* not happen anywhere in production and
is a config bug if it does.

This also helps with applying a partial config during unit tests on
actual hardware and not having a bunch of ports up and forwarding.

Reviewed By: ninas

Differential Revision: D4722038

fbshipit-source-id: 5bc5fb5e034ea46a4755629198040924fad7b018