ensure all hw ports are in SwitchState
Summary:
This modifies some of the setup apis in TestUtils to not try
to remove ports that are in our initial state if running on actual
hardware. This causes issues when running tests on actual hw since we
may get a link state notification for a port that does not exist in
the SwitchState.

Differential Revision: D4722082

fbshipit-source-id: 2787b85896eb99ed747f558b9ea82b6cfc642834