Add counters for v4, v6 and cumulative destination lookup failures
Summary:
We now trap packets for which a route could not be found
to (low pri queue on) CPU. This diff just adds counters for it.
In the layers we operate we should always have a default route.
So if DLF counters increment its a pretty high signal indication
that something is seriously wrong with routing.

Like other stats, have these be a cumalative count at the switch
level. We still go via the Port stats, so if desired we can also
build more granular counters to show which ports the DLF triggrring
packets came on.

NOTE that the low-pri queue is heavily policed, so we would get only
a subset of packets for which there were DLF failure. But the idea
here is to have a sustained non-zero count point us to routing failures.

Reviewed By: tfangit

Differential Revision: D4783202

fbshipit-source-id: 5661d40ddfbd422e841fe686a2560af323f377ae