Don't forgot port name when reseting to default state
Summary:
Similar to ID, name is a port identifier. When reseting to
default config state, retain name like we retain the port id.

Rename the API to better reflect what is done. The original intention
of the API also seems to be to just revert to default settings, while
retaining identifier(s).

This is not a problem on real switches, where all ports are present in
config. However, when testing with fake SDK, bcm_test code works with
sparse port tables and configs. This causes spurious updates for unanticipated
name changes.

In time we may make bcm_test to do platform specific initialization and
configs. But regardless of that, preventing spurious port name changes
seem the right thing to do here.

Reviewed By: plapukhov

Differential Revision: D4830453

fbshipit-source-id: 6a3fef4ce65a7c91a4e997e9231bf45e9f17d3d9