don't try to change speed on ports admin disabled in config
Summary:
We were hitting issues trying to change the speed on disabled
sub-ports when flexports are enabled. This is not supported in the sdk
and serves no real purpose outside of imposing more restrictions on
legal configurations.

This changes our BcmSwitch implementation to ignore changes to ports that
are disabled in software. If they ever get enabled, the changes will get
picked up then.

Differential Revision: D4869883

fbshipit-source-id: b3bfb23845d269729e2c0b884613a354d3f33ea2