Report packets discarded due to egress congestion
Summary:
Read the counter for packets dropped in CoS queue 0 on
egress. With mmu_lossless=0 egress is where all our congestion
drops should occur. And since we just one CoS queue just read
the counter for that.
When we start using more CoS queues - t16579288, the plan is to
have that be controlled via config. And since we will need to
configure these CoS queues per port, when that happens, we will
know which CoS queues to read stats for.

Differential Revision: D4804508

fbshipit-source-id: 17dfa52decac96862d768c2b0a5e286d9eea4a4b