Always call qsfp_service when configuring speed
Summary:
So, this is a bit messy. But, basically, because it's possible for a port to be up AND have incorrect CDR+rate select set, need to not make any assumptions about a port having the correct settings if it's up.
This removes a lot of the previous checks which would only reconfigure if speed changed. This will now also reconfigure on state changes, at warmboot time, when applying new config, and so on. This handles the case of a transceiver being removed and plugged back in.
The new code will only check whether a transceiver should be enabled before customizing it.

Differential Revision: D4650210

fbshipit-source-id: 08a14fff905409388d2637eb03056b1c3dc4a283