Fixup most of our fboss-oss build test issues
Summary:
The fboss-oss test was broken.  It turns out this had nothing to do with our code (it was an upstream thrift issue), but I did find and fix a bunch of problems with getdep.sh.  I also unified how legocastle testing is invoked.

I was not able to get legocastle to do full OSS builds as described in BUILD.md, which I will fork out to another task.
Legocastle testing:
* now facebook/lego/all-job-specs.sh returns an array of all legocastle testing jobs on each new diff
* incorporated the dependency testing into this new framework

Improvements to getdeps.sh:
* fixed the iproute2 git site to new (existing) location
* added two dependencies needed for 16.04
* added extensive error checking so we don't fail silently
* split out the build rules so we don't get confused
  * fbthrift now builds properly
* Added zstd and mstch info to CFLAGS and LDFLAGS

Reviewed By: ninas

Differential Revision: D4806505

fbshipit-source-id: 6e622027fe153b042805c791f00b05625b0c55b2