Remove redundant logic in route programming APIs
Summary:
In FBOSS thrift handler we support following operations
- AddRoute
- AddRoutes
- DelRoute
- DelRoutes
- SyncRoutes

and we implement each of these separately however we can leverage AddRoutes code
for AddRoute API (it doesn't have to be different). Since I am adding a small
enhancement to handle link-local nexthops, I thought of making this change so
that we have to maintain less code without compromising with APIs exposed by
FbossAgent.

Reviewed By: tfangit

Differential Revision: D4910892

fbshipit-source-id: 7fbbe3dc019c77b8685d1aa572bd16cfdca86062