Remove backward compatiility hacks for multiple-client feature
Summary:
In January-2017 pmonaco added support for multiple clients in FBOSS. This
change required some backward compatibility hacks. As of now this change has
been rolled out to whole fleet (being present in v118) so we can safely removes
these backward compatibility hacks.

Reason: I am making changes in route programming to support link-local nexthops.
It requires to touch this piece of code, so I am thinking it is better to get
rid of it now instead of letting it linger forever :)

Differential Revision: D4920617

fbshipit-source-id: 844245e4567d2f6e5d9e9a6b6051f5444536302a