LinkLocal Nexthop Programmig - Part1
Summary:
In this diff and an upcoming one, I am trying to add support for programming of
link local nexthops which requires scoping per interface. If a link local
nexthop is specified for programming it must have a scope specified.

This diff is only adding optional attribute in `RouteNextHop` which is set to
folly::none by default. Major change is in `RouteTypes.h` and all other changes
are just to make compilation happy. Since we are not really using `interface`
attribute of `RouteNextHop` it will not have any side-effect. Next diff will
make use of this attribute for link-local nexthops.

(Note: this ignores all push blocking failures!)

Reviewed By: tfangit

Differential Revision: D4911930

fbshipit-source-id: e000643ed98a9549a9c6b2bb84b620fddd0c2abf