Communicate initial speed to PlatformPort
Summary:
On warm boots, we do not need to change the port speed, so
never call linkSpeedChanged. The behavior I am seeing is that the
speed_ member var of WedgePort is not getting set for up ports on
warmboot. If we have ports that are down, we then see messages in the
log like "Unresolved speed: Unable to determine what qsfp
settings". If ports do go down, we would not customize properly
because we have no cached speed. This also means that if a port does
go down, we won't know how to properly customize it, breaking the
ability for some types of ports to come back after reseating the transceiver.

Longer-term we need to clean up the PortState semantics, which are
pretty confusing right now.

Differential Revision: D4874277

fbshipit-source-id: 1a6bd6a2795cb88f0df59bafadd9c9e477ba3384