Send pending entries to CPU rather then drop in h/w
Summary:
With the current behavior we set pending host entries
(for which we are resolving ARP/NDP) to drop. This causes
traffic to them to get discarded at ingress on other ports and
show up as in-discards in all our counters. This leads to spurious
debugging sessions, only to find that traffic being dropped is
due to these programmed drop entries.

So instead of dropping these packets, have them sent to CPU.
Elsewhere we already take care of sending them to the lo-pri
queue on CPU so it does not affect protocol traffic, which is
sent to hi-pri (BGP), mid-pri (ARP, NDP) queues.

This does attract more traffic to the lo-pri CPU queue, so
packets triggerring neighbor resolution have to compete with
other packets for destinations which had host entries set to
pending. Testing did not show this to be a problem.

Reviewed By: aeckert

Differential Revision: D4984211

fbshipit-source-id: edbc79df967a292b0b4ea6d9c7565dab43987f21