only customize transceivers on link status changes
Summary:
We had two issues where we were over-aggressively customizing
transceivers on warm boot.

The first was that we customized any enabled port in
linkStatusChanged. This meant that we would change transceiver
settings on up ports, which can cause a flap.

The second was in linkSpeedChanged, where we were unconditionally
customizing, hitting the same issue.

This diff changes the logic to only ever customize on link status
changes. We always call this after changing any port settings in
BcmPort and this will prevent duplicate customize calls when we
change the speed and then enable in one state update.

Differential Revision: D5021721

fbshipit-source-id: ecbbedae22655c82b85a762a8a05ae22bbf346dc