Add default route on startup when coldbooting for ALPM
Summary: ALPM requires that the first routes programmed be the defaults. Add static default routes with action drop. These then get overwritten with the correct next hops when the actual default route is added. Only do this on coldboot (warm boot means we already have programmed routes) and if ALPM is enabled.

Reviewed By: tfangit

Differential Revision: D4949302

fbshipit-source-id: 254e3016ce9d3cf54f630efc2c1473b3898e5269