Add AggregatePort and AggregatePortMap to SwitchState.
Summary:
This adds two new nodes to SwitchState: AggregatePort and AggregatePortMap.

An AggregatePort represents an IEEE 802.1AX aggregate port. That is, an AggregatePort represents a set of (physical) ports which behaves as if it were a single port with respect to a MAC client. AggregatePortMap holds the set of AggregatePorts in the system.

Further, the configuration format is exended to support configuring aggregate ports, and for ThriftConfigApplier to apply this configuration accordingly.

Note that if no aggregate ports are configured, the agent remains almost entirely unchanged in its dynamics (ie. the same code will be executed).

Reviewed By: tfangit

Differential Revision: D3737018

fbshipit-source-id: 1ea79523949efd26531fc7551a0a25dd52f0ce1f