Add BcmTrunk and BcmTrunkTable to BcmSwitch.
Summary:
This adds two new objects to BcmSwitch: BcmTrunk and BcmTrunkTable.

BcmTrunk encapsulates the logic to translate an AggregatePort to a Broadcom trunk port. BcmTrunkTable holds the set of BcmTrunk objects in the system.

In addition, the NodeMapDelta class is instantitated for AggregatePortMap to enable state deltas for aggregate ports.

Note that if link aggregation is not enabled (via the corresponding command-line flag), the modifications herein leave the state of the SDK, and consequently the ASIC, intact.

Reviewed By: tfangit

Differential Revision: D3818517

fbshipit-source-id: cf60275bb4d107f27ac47f91e625ebb1f313dc5f