Change call to qsfp service to be async for customizeTransceiver
Summary:
Use an async call instead of a sync call for customizeTransceivers. Do this on the background thread so that we don't block the update thread.
This change is needed because otherwise on hosts with many down transceivers we essentially get stuck behind a bunch of synchronous calls to the qsfp service, which in turn requires a lock on the pci bus for each transceiver. This can take upwards of 10s depending on the number of ports.
We don't actually need confirmation that this call succeeded before continueing, so just do it asynchronously.

Reviewed By: boryas

Differential Revision: D5100802

fbshipit-source-id: 1ec25fb8b497b9818a7390757d35a670bb698e36