Somewhat hacky way of not reporting stats twice for each port
Summary:
There is a short window b/w when BcmPorts objects
are created and when config get applied. Port names
come via config.
Reporting stats, for both named and unnamed
ports creates double the counters/stats.
In FB switch configs, we should always have port names,
so wait until port actually gets a name before
creating/reporting stats for it.

Considering the example in Test plan, # of keys reduced
per wedge40 is 1597 and those on wedge100 are 6272. There
are 11328 w40 and 9837 w100s deployed
P57390710

Taking rough numbers and assuming 10k each of w40 and w100
and rounding keys reduced to 1500 and 6000 respectively,
we should get a reduction of 75M with the next agent push.

Reviewed By: tfangit

Differential Revision: D5040159

fbshipit-source-id: 23e16136d5c527efd7b2573229895b8b5e0496cc